src/shared/requirements.txt
architect-functions
requests